+++ LightCTR/train/train_nfm_algo.h
@@ -48,14 +48,14 @@ private:
     size_t learnable_params_cnt;
     
     void batchGradCompute(size_t, size_t);
-    void accumWideGrad(size_t, double);
-    void accumDeepGrad(size_t, vector<double>*);
+    void accumWideGrad(size_t, float);
+    void accumDeepGrad(size_t, vector<float>*);
     
-    double *update_g;
-    inline double* update_W(size_t fid) {
+    float *update_g;
+    inline float* update_W(size_t fid) {
         return &update_g[fid];
     }
-    inline double* update_V(size_t fid, size_t facid) {
+    inline float* update_V(size_t fid, size_t facid) {
         assert(this->feature_cnt + fid * this->factor_cnt + facid <
                this->feature_cnt * (this->factor_cnt + 1));
         return &update_g[this->feature_cnt + fid * this->factor_cnt + facid];
